+++ ron-test/lib/StructSet/Types.hs
@@ -9,12 +9,11 @@ module StructSet.Types where
 
 import           RON.Prelude
 
-import           Data.Default (Default)
 import           RON.Schema.TH (mkReplicated)
 
 [mkReplicated|
     (struct_set StructSet13
-        int1 Integer    #ron{merge min}
+        int1 Integer    #ron{merge max}
         str2 RgaString
         str3 String     #ron{merge LWW}
         set4 (ORSet StructSet13)
@@ -22,7 +21,6 @@ import           RON.Schema.TH (mkReplicated)
         nst6 StructSet13)
 |]
 
-deriving instance Default StructSet13
 deriving instance Eq      StructSet13
 deriving instance Generic StructSet13
 deriving instance Show    StructSet13